ai_utility_suite/requirements.txt
# --- Core Python / System ---
ipykernel
python-dotenv
streamlit
fastapi[standard]

# --- LangChain & AI Core ---
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-google-community[gmail]==2.0.10
langchain-googledrive==0.3.35
langchain-groq==0.3.7
langchain-openai==0.3.30
langchain-text-splitters==0.3.11
langgraph==0.6.5
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.14

# --- Vector & Similarity Search ---
faiss-cpu
rapidfuzz

# --- Document Processing ---
pypdf
pdfplumber
azure-ai-documentintelligence

# --- Audio & Speech (TTS/STT) ---
pyttsx3
gTTS
edge-tts
streamlit-TTS
streamlit-mic-recorder

# --- Databases ---
psycopg2-binary

# --- Misc Utilities ---
dotenv
openai
